feat: support webpack 5 and rspack (#2)
Adds support for webpack 5 mainly by following the Webpack upgrade
documentation as well as mirroring a mix of the worker-rspack-loader
package which also had to go through the same upgrade.

See: https://github.com/rspack-contrib/worker-rspack-loader

As part of this I also upgraded most of the depedencies as it was easy
to do and very very out of date. Notably, did the following:

* Removed `del` and `del-cli`. Requires this to be an ESM module which
  wasn't worth it adn we don't really need these. Replaced with native
  node `fs` module.
* `loader-utilts` v3 doesn't work, so had to keep at v2.
* `eslint` v9 requires flat config which the
  `@webpack-contrib/eslint-config-webpack` package doesn't support